requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2020.6.20
Django==3.0.3
django-appconf==1.0.4
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==1.1.1
django-cities-light==3.6.0
django-debug-toolbar==2.2
django-filter @ file:///home/conda/feedstock_root/build_artifacts/django-filter_1591380509061/work
django-redis==4.12.1
django-select2==7.4.2
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1596632844333/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446511143/work
Markdown @ file:///C:/ci/markdown_1597415206643/work
olefile==0.46
Pillow @ file:///D:/bld/pillow_1594213067156/work
progressbar2==3.52.1
python-utils==2.4.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
wincertstore==0.2
zipp==3.1.0
